DriftingDroids include: minor changes
- remove unnecessary type conversions
- cast integer variables to long that could lead to integer overflow
- simplify for loop in Solver

+++ app/src/main/java/driftingdroids/model/KeyTrieSet.java
@@ -335,10 +335,10 @@ public final class KeyTrieSet {
     public final long getBytesAllocated() {
         long result = 0;
         for (int i = 0;  i < this.numNodeArrays;  ++i) {
-            result += this.nodeArrays[i].length << 2;
+            result += (long) this.nodeArrays[i].length << 2;
         }
         for (int i = 0;  i < this.numLeafArrays;  ++i) {
-            result += this.leafArrays[i].length << 2;
+            result += (long) this.leafArrays[i].length << 2;
         }
         return result;
     }
